+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-iap",
-  "version": "6.3.0",
+  "version": "7.0.0",
   "description": "React Native In App Purchase Module.",
   "main": "index.js",
   "types": "index.d.ts",
